Creating initial commit
Version 0 of tidy_data for AV dual target paradigm. This project
converts from the cumbersome beethoven data files to a simpler and more
straightforward to interpret table of data. Currently this code is only
set up to work the with most up to date paradigm version (AVD2) but will
expand in the future to be backwards compatible with the AVD1 paradigm.